==> RecSysProj/lab12_ts/01 basic predictors.ipynb <==
if True:
    import os
    os.environ["PYSPARK_SUBMIT_ARGS"]='--packages com.databricks:spark-csv_2.10:1.3.0 pyspark-shell'
    execfile(os.path.join(os.environ["SPARK_HOME"], 'python/pyspark/shell.py'))
    import os
    import sys
else:
    os.environ['SPARK_HOME'] = '/usr/lib/spark'
    sys.path.insert(0, '/usr/lib/spark/python/lib/py4j-0.9-src.zip')
    sys.path.insert(0, '/usr/lib/spark/python/')
    sys.path.insert(0, '/usr/local/lib64/python2.7/site-packages')
    sys.path.insert(0,'/usr/local/lib/python2.7/site-packages')
    
    from pyspark import SparkContext
    from pyspark.sql import SQLContext, HiveContext
    
    try: sc = SparkContext()
    except: None    
    sqlc = SQLContext(sc)
    spark = sqlc

from pyspark.sql.types import StructType, StructField, StringType, IntegerType, FloatType, MapType
import pyspark.sql.functions as F

%pylab inline

#spark = sqlc

import seaborn as sns

import numpy as np
import pandas as pd
import time
import json

from operator import add
from pyspark.sql.types import *

from pyspark.mllib.recommendation import Rating
from pyspark.mllib.recommendation import ALS, MatrixFactorizationModel
from datetime import datetime

def convert_uir_tupple(line):
    r = line.split(',')
    return (int(r[0]), int(r[1]), float(r[2]))

def load_train():
    raw = sc.textFile("lab_12/input/train.csv")\
            .filter(lambda x: not x.startswith("userId"))\
            .map(convert_uir_tupple)
    return raw.toDF(schema=StructType([StructField("userId",  IntegerType()),
                                           StructField("movieId", IntegerType()),
                                           StructField("rating",  FloatType())])) 
    

train = load_train()

train.count()

# rerun eval notebook with full data set

strain = train

# Baseline with Basic predictiors approach

smooth_buser = 5
smooth_bitem = 10

avg_rating = strain.agg({"rating":"mean"}).collect()[0]

avg_rating = avg_rating[0]
print 'avg rating', avg_rating

buser = strain.groupBy("userId").agg(F.sum(F.col("rating")).alias("sum_ratings"), 
                                     F.count(F.col("rating")).alias("n_user_ratings"))

buser.show(2)

x = buser.select("userId", "n_user_ratings", 
                 ((F.col("sum_ratings") - F.lit(avg_rating)*F.col("n_user_ratings"))  / (F.col("n_user_ratings") + F.lit(smooth_buser))).alias("bpu"))

x.show(2)

buser = x

strain.registerTempTable("strain")

buser.registerTempTable("buser")

bitem = spark.sql("""
            select
                t.movieId, sum(t.rating - %g - bu.bpu) as sum_r_normalised, count(*) n_item_ratings
            from strain t
            join buser bu on t.userId = bu.userId 
            group by t.movieId
""" % (avg_rating,))

bitem.show(2)

x = bitem.select("movieId", "n_item_ratings", 
                 ( F.col("sum_r_normalised")  / (F.col("n_item_ratings") + F.lit(smooth_bitem))).alias("bpi"))

x.show(2)

bitem = x

bitem.registerTempTable("bitem")

## Build prediction for strain set

q = spark.sql("""
        select 
            t.*,
            bu.n_user_ratings,
            bi.n_item_ratings,
            ({avg_rating} + bu.bpu + bi.bpi) as bp_predicted_rating,
            rating - ({avg_rating} + bu.bpu + bi.bpi) as bp_err
        from strain t
        join buser bu on t.userId = bu.userId 
        join bitem bi on t.movieId = bi.movieId
""".format(avg_rating=avg_rating))

strain_predictions = q

strain_predictions.show(2)

def calc_rmse(df, prediction_col_name):
    mse = df.select(F.pow(F.col(prediction_col_name) - F.col("rating"), F.lit(2)).alias("err"))\
            .agg({"err":"mean"}).collect()[0][0]
    return np.sqrt(mse)    

rmse = calc_rmse(strain_predictions, "bp_predicted_rating")

print 'strain RMSE', rmse

def cut_1to5(x):
    if x < 1.:
        return 1.0
    if x > 5:
        return 5.0
    return x

udf_cut_1to5 = F.udf(cut_1to5, FloatType())

#q = strain_predictions.withColumn("predicted_rating", udf_cut_1to5(F.col("predicted_rating")))

#q.show(2)

#print 'strain RMSE with cut1t5', calc_rmse(q)

## Cache results

#strain_predictions.coalesce(8).write.parquet("lab_12/cache/full_train_bp",
#                                             mode='overwrite')

## Build prediction for test set

def convert_ui_tupple(line):
    r = line.split(',')
    return (int(r[0]), int(r[1]))
    
def load_test():    
    raw = sc.textFile("lab_12/input/test.csv")\
            .filter(lambda x: not x.startswith("userId"))\
            .map(convert_ui_tupple)
    return raw.toDF(schema=StructType([StructField("userId",  IntegerType()),
                                       StructField("movieId", IntegerType())]))

test = load_test()

test.count()

stest = test

stest.registerTempTable("stest")

q = spark.sql("""
        select 
            t.*,
            bu.n_user_ratings,
            bi.n_item_ratings,
            (%g + coalesce(bu.bpu, 0.0) + coalesce(bi.bpi, 0.0)) as bp_predicted_rating
        from stest t
        left join buser bu on t.userId = bu.userId 
        left join bitem bi on t.movieId = bi.movieId
""" % (avg_rating, ))

q.show(2)

stest_predictions = q

## Cache bp predictions on full test

#stest_predictions.coalesce(8).write.parquet("lab_12/cache/full_test_bp",
#                                             mode='overwrite')

def cut_1to5(x):
    if x < 1.:
        return 1.0
    if x > 5:
        return 5.0
    return x

udf_cut_1to5 = F.udf(cut_1to5, FloatType())


qs = q.select("userId", "movieId", udf_cut_1to5(F.col("predicted_rating")).alias("rating")).orderBy("userId", "movieId")

res = qs.toPandas()

res.to_csv("lab_12.csv", index=False)
