==> RecSysProj/lab12_ts/13 als predict stest on full score.ipynb <==
if True:
    import os
    os.environ["PYSPARK_SUBMIT_ARGS"]='--packages com.databricks:spark-csv_2.10:1.3.0 pyspark-shell'
    execfile(os.path.join(os.environ["SPARK_HOME"], 'python/pyspark/shell.py'))
    import os
    import sys
else:
    os.environ['SPARK_HOME'] = '/usr/lib/spark'
    sys.path.insert(0, '/usr/lib/spark/python/lib/py4j-0.9-src.zip')
    sys.path.insert(0, '/usr/lib/spark/python/')
    sys.path.insert(0, '/usr/local/lib64/python2.7/site-packages')
    sys.path.insert(0,'/usr/local/lib/python2.7/site-packages')
    
    from pyspark import SparkContext
    from pyspark.sql import SQLContext, HiveContext
    
    try: sc = SparkContext()
    except: None    
    sqlc = SQLContext(sc)
    spark = sqlc

%pylab inline

import seaborn as sns

import numpy as np
import pandas as pd
import time
import json

from pyspark.sql.types import StructType, StructField, StringType, IntegerType, FloatType, MapType
import pyspark.sql.functions as F

from operator import add
from pyspark.sql.types import *

from pyspark.mllib.recommendation import Rating
from pyspark.mllib.recommendation import ALS, MatrixFactorizationModel
from datetime import datetime

# Load data

strain_predictions = spark.read.parquet("lab_12/cache/strain_bp")


strain_predictions.show(2)

stest_predictions = spark.read.parquet("lab_12/cache/stest_bp")

stest_predictions.show(2)

# ALS on full score

## Prepeare train,validate,test data

strain_predictions.rdd.take(1)

def convert_to_als_rating(r):
    return Rating(r.userId, r.movieId, r.rating)

strain_ratings = strain_predictions.rdd.map(convert_to_als_rating)

strain_ratings.first()

# Train

rank = 25
numIterations = 15
lambda_ = 0.1

%%time
model = ALS.train(strain_ratings, rank, numIterations, lambda_=lambda_, seed=68)

# Predict on test

stest_raw = model.predictAll(stest_predictions.rdd.map(lambda x: (x.userId, x.movieId)))

stest_raw.first()

stest_als_pred = stest_raw.toDF(schema=StructType([
                                            StructField("userId",  IntegerType()),
                                            StructField("movieId", IntegerType()),
                                            StructField("predicted_als_rating",  FloatType())]))

stest_als_pred.show(3)

#stest_als_pred.coalesce(8).write.parquet("lab_12/cache/stest_als25_has_missed", 
#                                         mode='overwrite')

# Eval

stest_predictions = spark.read.parquet("lab_12/cache/stest_bp_i2i")

stest_predictions.columns

stest_predictions.count()

eval_data = stest_predictions.join(stest_als_pred, on=['userId', 'movieId'], how='left')

eval_data.cache()

def rmse(df, prediction_column):
    mse_df=df.select(F.pow(F.col("rating")-F.col(prediction_column), 2).alias("e2"))\
             .agg(F.avg(F.col("e2")).alias("mse"),
                  F.count(F.col("e2")).alias("n_records"))
    z=mse_df.collect()[0]
    return np.sqrt(z[0]), z[1]

z = rmse(eval_data, "predicted_als_rating")

# (0.83440097064385499, 3154183)
print z

#top50
#i2i on >2 (0.77253385132439278, 759878)
#als on >2 (0.74363847834562224, 758677)
n=2
z = eval_data.filter(F.col("n_used_item_ratings")>n)
print 'i2i on >',n, rmse(z, "predicted_i2i_rating")
print 'als on >',n, rmse(z, "predicted_als_rating")


eval_data=eval_data.withColumn("predicted_rating", F.coalesce(F.col("predicted_als_rating"), F.col("bp_rating")))

eval_data=eval_data.withColumn("predicted_rating_avg",
                               F.lit(0.75) * F.col("predicted_rating") +  F.lit(0.25) * F.col("predicted_i2i_rating"))

eval_data.cache()

n=0
z = eval_data.filter(F.col("n_used_item_ratings")>=n)
#print 'bp on >=',n, rmse(z, "bp_rating")
#print 'i2i on >=',n, rmse(z, "predicted_i2i_rating")
print 'als on >=',n, rmse(z, "predicted_rating")
print 'avg on >=',n, rmse(z, "predicted_rating_avg")

#top50
#als on >= 0 (0.83495897716125911, 3159915)
#avg on >= 0 (0.8236655014538885, 3159915)
#top100
#als on >= 0 (0.83495897715420975, 3159915)
#avg on >= 0 (0.82292474871365584, 3159915)

def get_rmse_per_category(eval_data, prediction_column):
    qg = eval_data.withColumn("e2", F.pow(F.col("rating")-F.col(prediction_column), 2))\
             .groupBy(["nr_user_cat", "nr_item_cat"])\
             .agg(F.avg(F.col("e2")).alias("mse"), 
                  F.count(F.col("e2")).alias("n_examples"),
                  F.sum(F.col("has_i2i_rating")).alias("has_i2i_rating"),
                  F.avg(F.col("n_used_item_ratings")).alias("avg_used_item_ratings"))
    epg = qg.toPandas()
    epg['r_examples'] = np.round(epg['n_examples'] * 100. /  epg['n_examples'].sum(), 2)
    epg['rmse'] = np.sqrt(epg['mse'])
    epg['r_has_i2i_rating'] = np.round(epg['has_i2i_rating'] * 100. /  epg['n_examples'], 2)
    
    return epg[["nr_user_cat", "nr_item_cat", 
                "n_examples", "avg_used_item_ratings",
                "r_examples", "rmse", "r_has_i2i_rating"]].sort_values(by='r_examples', ascending=False)
    return epg

get_rmse_per_category(eval_data, "predicted_rating").head(4)

get_rmse_per_category(eval_data, "predicted_rating_avg").head(4)
