==> RecSysProj/lab12_ts/15 als eval u2u.ipynb <==
if True:
    import os
    os.environ["PYSPARK_SUBMIT_ARGS"]='--packages com.databricks:spark-csv_2.10:1.3.0 pyspark-shell'
    execfile(os.path.join(os.environ["SPARK_HOME"], 'python/pyspark/shell.py'))
    import os
    import sys
else:
    os.environ['SPARK_HOME'] = '/usr/lib/spark'
    sys.path.insert(0, '/usr/lib/spark/python/lib/py4j-0.9-src.zip')
    sys.path.insert(0, '/usr/lib/spark/python/')
    sys.path.insert(0, '/usr/local/lib64/python2.7/site-packages')
    sys.path.insert(0,'/usr/local/lib/python2.7/site-packages')
    
    from pyspark import SparkContext
    from pyspark.sql import SQLContext, HiveContext
    
    try: sc = SparkContext()
    except: None    
    sqlc = SQLContext(sc)
    spark = sqlc

%pylab inline

import seaborn as sns

import numpy as np
import pandas as pd
import time
import json

from pyspark.sql.types import StructType, StructField, StringType, IntegerType, FloatType, MapType
import pyspark.sql.functions as F

from operator import add
from pyspark.sql.types import *

from pyspark.mllib.recommendation import Rating
from pyspark.mllib.recommendation import ALS, MatrixFactorizationModel
from datetime import datetime

# Load data

strain_predictions = spark.read.parquet("lab_12/cache/strain_bp_v1")

strain_predictions.show(2)

stest_predictions = spark.read.parquet("lab_12/cache/stest_bp_v1")

stest_predictions.show(2)

# Calculate normalised user rating per movie

avg_rating = strain_predictions.agg({"rating":"mean"}).collect()[0][0]
print "avg_rating", avg_rating

strain_predictions = strain_predictions.withColumn("user_avg_rating", 
                        (F.col("user_sum_ratings") + F.lit(avg_rating)) / (F.col("n_user_ratings") + F.lit(1.0)))
strain_predictions = strain_predictions.withColumn("user_norm_rating", F.col("rating") - F.col("user_avg_rating"))
strain_predictions = strain_predictions.select("userId", "movieId", "user_norm_rating").cache()

strain_predictions.show(2)

strain_predictions.registerTempTable("strain")

stest_predictions = stest_predictions.withColumn("user_avg_rating", 
                        (F.col("user_sum_ratings") + F.lit(avg_rating)) / (F.col("n_user_ratings") + F.lit(1.0)))
stest_predictions = stest_predictions.select("userId", "movieId", "rating", 
                                             "user_avg_rating", "nr_user_cat", "nr_item_cat", "bp_rating").cache()

stest_predictions.show(2)

stest_predictions.registerTempTable("stest")

#stest_predictions.cache()

# Load similarity matrix

def convert_u2u_tupple(line):
    r = line.split(',')
    return (int(r[0]), int(r[1]), float(r[2]), int(r[3]))

def load_u2u_sim():
    raw = sc.textFile("lab_12/u2u_top100to200.csv")\
            .filter(lambda x: not x.startswith("user"))\
            .map(convert_u2u_tupple)
    return raw.toDF(schema=StructType([StructField("user1",  IntegerType()),
                                       StructField("user2", IntegerType()),
                                       StructField("sim",  FloatType()),
                                       StructField("sim_rank",  IntegerType())])) 
    return raw

u2u = load_u2u_sim()

u2u.show(2)

u2u.cache()

u2u.registerTempTable("u2u_sim")

strain_predictions.dtypes

# Make predictions on test

q = spark.sql("""
        select
          pr.userId,
          pr.movieId,
          avg(
              if(pr.user_avg_rating < 0.001, cast({avg_rating} as double), pr.user_avg_rating)
          ) + 
              sum(kr.user_norm_rating * s.sim) /  sum(s.sim)  as predicted_rating,
          count(*) as n_used_ratings
        from stest pr
        join u2u_sim s on pr.userId = s.user1 and s.sim_rank <= 150
        join strain kr on pr.movieId = kr.movieId and kr.userId = s.user2
        group by pr.userId, pr.movieId
""".format(avg_rating=avg_rating))

q.show(2)

predicted_on_test = q

predicted_on_test.registerTempTable("predicted_test")

eval_data = spark.sql("""
        select
          tbase.*,
          coalesce(pt.n_used_ratings, 0) as n_used_user_ratings,
          coalesce(predicted_rating, bp_rating) as predicted_rating,
          if(pt.n_used_ratings is not null, 1, 0) as has_u2u_rating
        from stest tbase
        left join predicted_test  pt on tbase.userId = pt.userId and tbase.movieId = pt.movieId
""")

eval_data.cache()

eval_data.show(2)

msedf=eval_data.select(F.pow(F.col("rating")-F.col("predicted_rating"), 2).alias("e2"))\
               .agg(F.avg(F.col("e2")).alias("mse"))

print 'rmse on all test set:', np.sqrt(msedf.collect()[0][0])

msedf_u2uonly=eval_data.filter(F.col("has_u2u_rating")==1) \
                       .select(F.pow(F.col("rating")-F.col("predicted_rating"), 2).alias("e2"))\
                       .agg(F.avg(F.col("e2")).alias("mse"))

print 'rmse on records with u2u predictions:', np.sqrt(msedf_u2uonly.collect()[0][0])

msedf_u2uonly_t3=eval_data.filter(F.col("n_used_user_ratings")>=3) \
                       .select(F.pow(F.col("rating")-F.col("predicted_rating"), 2).alias("e2"))\
                       .agg(F.avg(F.col("e2")).alias("mse"))

print 'rmse on records with at least 3 similar u2u predictions:', np.sqrt(msedf_u2uonly_t3.collect()[0][0])

qg = eval_data.withColumn("e2", F.pow(F.col("rating")-F.col("predicted_rating"), 2))\
             .groupBy(["nr_user_cat", "nr_item_cat"])\
             .agg(F.avg(F.col("e2")).alias("mse"), 
                  F.count(F.col("e2")).alias("n_examples"),
                  F.sum(F.col("has_u2u_rating")).alias("has_u2u_rating"),
                  F.avg(F.col("n_used_user_ratings")).alias("avg_used_user_ratings"))

epg = qg.toPandas()

epg['r_examples'] = np.round(epg['n_examples'] * 100. /  epg['n_examples'].sum(), 2)
epg['rmse'] = np.sqrt(epg['mse'])
epg['r_has_u2u_rating'] = np.round(epg['has_u2u_rating'] * 100. /  epg['n_examples'], 2)

epg.sort_values(by='r_examples', ascending=False)

print eval_data.filter(F.col("n_used_user_ratings")>=3).count()
print eval_data.count()

eval_data.select("userId", "movieId", "rating", "nr_user_cat", "nr_item_cat", 
                 "n_used_user_ratings", "has_u2u_rating", 
                 F.col("predicted_rating").alias("predicted_u2u_rating")).show(2)

# Save for ensemble cheks

eval_data = eval_data.select("userId", "movieId", "rating", "nr_user_cat", "nr_item_cat", 
                 "n_used_user_ratings", "has_u2u_rating", 
                 F.col("predicted_rating").alias("predicted_u2u_rating"))

eval_data = eval_data.coalesce(8)

eval_data.show(2)

eval_data.write.parquet("lab_12/cache/stest_bp_u2u_top150", mode='overwrite')
