==> RecSysProj/lab12_ts/20 extract pivot for ensembling.ipynb <==
if True:
    import os
    os.environ["PYSPARK_SUBMIT_ARGS"]='--packages com.databricks:spark-csv_2.10:1.3.0 pyspark-shell'
    execfile(os.path.join(os.environ["SPARK_HOME"], 'python/pyspark/shell.py'))
    import os
    import sys
else:
    os.environ['SPARK_HOME'] = '/usr/lib/spark'
    sys.path.insert(0, '/usr/lib/spark/python/lib/py4j-0.9-src.zip')
    sys.path.insert(0, '/usr/lib/spark/python/')
    sys.path.insert(0, '/usr/local/lib64/python2.7/site-packages')
    sys.path.insert(0,'/usr/local/lib/python2.7/site-packages')
    
    from pyspark import SparkContext
    from pyspark.sql import SQLContext, HiveContext
    
    try: sc = SparkContext()
    except: None    
    sqlc = SQLContext(sc)
    spark = sqlc

%pylab inline

import seaborn as sns

import numpy as np
import pandas as pd
import time
import json

from pyspark.sql.types import StructType, StructField, StringType, IntegerType, FloatType, MapType
import pyspark.sql.functions as F

from operator import add
from pyspark.sql.types import *

from pyspark.mllib.recommendation import Rating
from pyspark.mllib.recommendation import ALS, MatrixFactorizationModel
from datetime import datetime

# Load data

df_i2i = spark.read.parquet("lab_12/cache/stest_bp_i2i/")

df_i2i.show(2)

df_i2i.count()

df_als = spark.read.parquet("lab_12/cache/stest_als25_has_missed")

df_als.show(2)

df_als.count()

df_u2u = spark.read.parquet("lab_12/cache/stest_bp_u2u_top150/")

df_u2u.show(2)

df_u2u.count()

# Merge everything in single pivot

df_i2i.registerTempTable("df_i2i")
df_als.registerTempTable("df_als")
df_u2u.registerTempTable("df_u2u")

df_i2i.columns

df_als.columns

df_u2u.columns

q = spark.sql("""
      select
          i.*,
          coalesce(u.n_used_user_ratings, 0.0) n_used_user_ratings,
          coalesce(u.has_u2u_rating, 0) has_u2u_rating,
          coalesce(u.predicted_u2u_rating, i.bp_rating) as predicted_u2u_rating,
          coalesce(a.predicted_als_rating, i.bp_rating) as predicted_als_rating,
          if(a.predicted_als_rating is not null, 1, 0) as has_als_rating
      from df_i2i i
      left join df_u2u u on i.userId = u.userId and i.movieId = u.movieId
      left join df_als a on i.userId = a.userId and i.movieId = a.movieId
      
""")

q.show(2)

q.cache()

q.count()

q = q.coalesce(8)

#q.write.parquet("lab_12/cache/stest_pivot4ensembling",  mode='overwrite')

#res = q.toPandas()

# Read than get pandas

q = spark.read.parquet("lab_12/cache/stest_pivot4ensembling")

q.show(2)

q.count()

qrdd = q.rdd.coalesce(30).cache()

def make_part_filter(index):
    def part_filter(split_index, iterator):
        if split_index == index:
            for el in iterator:
                yield el
    return part_filter


tmp_dir = "/data/home/taras.svirsky/lab12/tmp/"
if not os.path.exists(tmp_dir):
    os.makedirs(tmp_dir)

import cPickle as pickle

for part_id in range(qrdd.getNumPartitions()):
    part_rdd = qrdd.mapPartitionsWithIndex(make_part_filter(part_id), True)
    data_from_part_rdd = part_rdd.collect()
    print "partition id: %s elements: %s" % (part_id, data_from_part_rdd)
    print part_id, len(data_from_part_rdd),
    fpath = os.path.join(tmp_dir,"part%6d.pckl" % (part_id,))
    with open(fpath, 'w') as outh:
        pickle.dump(data_from_part_rdd, outh)
    print 'saved to', fpath    

2+2

asd=pd.DataFrame.from_records(data_from_part_rdd)

zz=data_from_part_rdd[0]

columns = q.columns

df_list = []
for f in os.listdir(tmp_dir):
    print f
    with open(os.path.join(tmp_dir, f)) as inh:
        rows = pickle.load(inh)
    sdf = pd.DataFrame.from_records(rows, columns=columns)
    df_list.append(sdf)

sum([len(x) for x in df_list])

df = pd.concat(df_list)

df.head()

df.to_csv("/data/home/taras.svirsky/lab12/data/stest_pivot4ensemble.csv", index=False)

df.shape
